Stop using symlinks for desktop file and icon, copy them instead
Two users recently reported GNOME Shell crashes. Although the exact cause
was unclear from the logs, one  user noticed the issue occurred when opening
Nautilus while this extension was enabled.

A similar problem was previously discussed in the GNOME Extensions Matrix channel.
Other developers experienced GNOME Shell crashes when their extensions were updated to a new version that used
symbolic links.

While it’s not confirmed that symlinks are the cause in this case, it’s a plausible factor.
To avoid gambling on stability, remove use of symlinks entirely